logrotate: upgrade to 3.8.7
This is a major upgrade from an almost ten years old release to the
current release. I reviewed the patch series against current Debian
(sid) patches and upstream changes. Everything not in the current Debian
patches went upstream. All patches gone upstream were discarded. Due to
the major changes, Git won't recognize each rename of the remaining
ones.

There was one old patch ptx-timeext.patch which seemed to be ptxdist
specifix, but it does not apply anymore. It was removed. If anyone needs
the features from it, maybe she could extract it from the previous state
and port it by herself.

Signed-off-by: Alexander Dahl <post@lespocky.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>