add ptxsudo alias to ptxdist bash
When using 'ptxdist bash' some tools installed by ptxdist may have to be
used as root. some sudo installations do not pass through the PATH environment
variable so that 'sudo <somecommand-in-sysroot-host>' results in command not
found error messages. This patch adds a ptxsudo alias that can be used in such
situations. The name has been chosen rather than regular 'sudo' since the
said feature is a security feature protecting against path spoofing which
we do not want to disable silently.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>