libusb: drop libusb and use libusbx
This patch drops the libusb support and add the libusbx instead.

We can do that because libusb and libusbx offers the same API, original
text from libusbx project website:

"If you are an existing user of libusb, you don't have to change
anything. Just download the libusbx library and use it in your code. The
library is designed as a drop in replacement for libusb, so please give
it a try. If you don't like it, which we very much doubt, you can always
revert..."

Another point is that libusbx website says that the libusb project is
somehow died and will not longer release new releases:

"The reason for the fork is that, despite having dedicated members,
libusb has still not been able to produce a new release for the past 2
years. When a project fails to produce regular releases, we consider
that you, its user, are paying the ultimate price. This is because it
means that patches and new feature are being witheld and you end up
wasting your time."

This patch fix also a few bugs which I detected:

i) switch to "--disable-static", otherwise it was "--enable-shared
   --enable-static"

ii) correct use of --disable-log and --enable-debug-log, there was an
   option for these two configure flags but also was a static
   "--disable-log --disable-debug-log" in the configure call, which
   makes no sense.

iii) Fix Jürgen Beisert's mail address.

iv) cleanup and remove _ENV and _PATH setting.

new features:

i) Add LIBUSB_SYSTEM_LOG which could enable --enable-system-log flag.

ii) also adding some new disable flag --disable-tests-build

iii) I add "--disable-udev" there but we should have a --enable-udev...
    The old one has no kind of udev support and I got a recursive
    dependency detection because udev-legacy needs usbutils which needs
    libusb. We should look at this in the near future.

iv) change "libusb" entry to "libusbx"

v) rename LIBUSB_DISBALE_LOG to LIBUSB_LOG and change logic.

I also keep the patch from libusb and port it to libusbx because I saw
that libusbx uses also pthreads.

Signed-off-by: Alexander Aring <alex.aring@gmail.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>