libcoap: add new package
This patch adds the libcoap implementation [0]. The most common use of
this library is to use it with contiki os. That's one reason why the
current buildsystem isn't "well". The added patches replace the current
buildsystem with a better one. The replacement had the goal to made less
changes to the current libcoap implementation.

The libcoap implementation do a include of autoconf "config.h" header
file in the API headers. The patches adds a new libcoap_config.h with a
LIBCOAP_ prefix to all defines and replace the use of these defines in
the API headers. That's too not well but offers a workaround to make a
clean install of this library and other autotools software can use it.

I found in the current release some things which are broken like some
disable features in configure call and others things which are not well.
I don't fix it in the patches, I will try to send the fixes to the
maintainer of this project and hopefully it's fixed in the next release.

Libcoap is still the best known C implementation of the coap protocol,
but also it's a bleeding edge implementation of a draft [1].

[0] http://sourceforge.net/projects/libcoap/
[1] http://tools.ietf.org/html/draft-ietf-core-coap-18

Signed-off-by: Alexander Aring <alex.aring@gmail.com>
[mol: simplify first patch]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>