cmatrix: add new package
This is a must have tool for linux hackers. This tool isn't
cross-compile friendly. The added patch removes the static filename
check for consolefonts path. This is removed and should handeld by a
select on KBD_CONSOLEFONTS.

Another issue is that "make install-data-local" should install fonts
into consolefonts directory. This is also a static path without a
${prefix} variable. This is solved by install the necessary fonts
manually in targetinstall. Currently there is no X11 font support.

Signed-off-by: Alexander Aring <alex.aring@gmail.com>
[mol: regenerate patches with git-ptx-patches]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>