newpackages: provide the M4 macro directory to the wizard script
In order to be able to keep all our cross compile M4 macros at a single
location and share it between all autotools based templates we must forward
the location of the M4 macros to the wizard script. This change adds a second
parameter when calling the template's wizard. Older wizards will ignore it so
it should not confuse existing templates.

Signed-off-by: Juergen Borleis <jbe@pengutronix.de>