ptxd_make_world_install_mangle_pc: use '${prefix}/lib' for libdir
The regex is basically <prefix>/lib so '${prefix}/lib' is the correct
value. Usually that is not a problem because prefix == exec_prefix, but
this fails for .pc files that do not define exec_prefix at all.

Without this checking for libpci fails with:
Variable 'exec_prefix' not defined in '.../sysroot-target/usr/lib/pkgconfig/libpci.pc'

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>