ptxd_lib_dgen: be more clever about tarball dependencies
New package versions have always been problematic. If the old version
is built and the tarball for the new version is older than the get
stage then ptxdist didn't build the new version.
To improve the situation we now do the following:
- create a dummy state file with the tarball name as part of the
  filename.
- let the get stage depend on this.
For a new version the file does not exist and the package is rebuilt.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>