ptxdist: Create symlinks for toolchain-related tools
This patch adds symlinks in the toolchain wrapper dir for
toolchain-related tools like as, nm, etc. Currently, that symlinks are
only created for tools for which ptxdist provides the wrapper mechanism.

The primary use case is our internal build system (application that
resides in local_src) that expects all programs to be at the same
location (because some very complicated magic...).

Apart from that, I think the solution to have all tools at the same
place is cleaner and the patch is very small. So it would be nice if it
could be accepted.

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>