archivemount: new package for archivemount 0.8.3
Archivemount is a piece of glue code between libarchive and FUSE.
It can be used to mount a (possibly compressed) archive (as
in .tar.gz or .tar.bz2) and use it like an ordinary filesystem.

It is a user-space application setting up the mount.
Usage: archivemount <archivepath> <mountpoint> [options]

For more information, see
http://sourceforge.net/p/fuse/wiki/ArchiveFileSystems/#archivemount-mount-archives-tar-cpio
http://www.cybernoia.de/software/archivemount.html

Signed-off-by: Andreas Pretzsch <apr@cn-eng.de>
[mol: small cosmetic changes]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>