GDB: switch to an available archive type for all releases.
The to be used release for GDB can be defined in the menu. Upstream now seems
to prefer the xz archive format for newer releases, but kept the bz2 for older
releases. Only the gz archive format is available for all releases.
By switching to the gz archive format should work for all releases now.
Drawback: gz archives are much larger than bz2 or xz archives.

Signed-off-by: Juergen Borleis <jbe@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>