pyserial3: Provide pyserial package for Python 3
This upstream package can be used for both Python 2.7 and Python 3.

I'm providing the python 3 variant as 'pyserial3'. The package is designed so
that it can be built and installed in parallel to pyserial (Python 2
installation). So I renamed the miniterm.py to miniterm3.py.

The other example are not installed as renaming all of them wasn't worth the
effort in my opinion.

Signed-off-by: Bernhard Walle <bernhard@bwalle.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>