ptxd_make_world_install_post: fix race condition
Just copying ${pkg_pkg_dir} recursively can fail. This can happen with -jeX
when 'cp' is executed for two packages the same time: 'cp' first checks if a
directory exists, and creates it if not. If two instances of 'cp' try this with
the same directory at the same time, there is a race condition and the second
'cp' may fail to create the directory because it already exists.

Avoid this by creating the directories first. 'mkdir -p' will just try to
create the directories and ignores any EEXIST.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>