bash: Updated the bash shell to release 4.3.30
systemd seems to use modern bash features like readarray, which is not
provided by version 3.2.48 on current ptxdist.

So I updated the bash package to a more recent bash version.

The config fragment was updated with actual configure arguments, so that bash
should be configurable similiar as for 3.2 version.

Removed unneeded dependency statements and other tweaks like unconditionally
selecting BASH_CONDITIONAL.

The patches for version 3.2.48 are dropped as these patches are no longer
needed for version 4.3. At least I got told so.

Signed-off-by: Carsten Schlote <schlote@vahanus.net>
[mol: add new and remove old patches, BASH_SEP_HELPFILES is broken]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>