image: ipkg, opkg: use symbol IMAGE_XPKG_EXTRA_ARGS to pass extra arguments to ipkg and opkg build
Originally the symbols IMAGE_IPKG_EXTRA_ARGS, IMAGE_OPKG_EXTRA_ARGS were used
in the code, but there only was the config entry for the ipkg symbol in
Kconfig.

This patch converts the IMAGE_IPKG_EXTRA_ARGS to IMAGE_XPKG_EXTRA_ARGS and uses
it during ipkg and opkg build.

Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>