openssl: keep older releases downloadable
If new bugfix releases are made, upstream moves the previous release to a
different location. Follow to this different location.

Note: this change might be a bad idea because it keeps older and vulnerable
releases in use. But it also helps to keep BSPs build able and not to fail
if the upstream archive disappears.

Signed-off-by: Juergen Borleis <jbe@pengutronix.de>