libtxc_dxtn: new package
This adds libtxc_dxtn, which is a free software implementation
of the S3TC texture compression. If this is enabled MESA will
provide the EXT_texture_compression_s3tc OpenGL extension.

It is a separate library, as S3TC is still patent encumbered
and might need some special attention when shipping products.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>