ptxd_make_world_patchin_apply_git: make sure the series file exists
The series file for 'git quiltimport' was only created if the original series
file is not empty. In this case, with newer git versions (at least 2.8.1)
'git quiltimport' returns an error if the series file is missing.

An empty series file is usefull to start creating new patches.
Create the series file for 'git quiltimport' to make sure this works with
new git versions.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>