mySQL: exhume a gone package by updating it to a new release
After the remove of the old package from the staging area this change adds a
new release of the mySQL package.

Its intended and checked on a read-only root filesystem with write access
in "/var/mysql" only. The database engine runs as a regular user and does
not need root permissions.

Due to its dependency to boost_1_59_0 it uses exactly this version and the
already present boost lib in PTXdist isn't used.

Be aware: its buildsystem uses a huge amount of memory. If your host is low
on memory don't forget to build this package with the '-ji1' parameter.

Signed-off-by: Juergen Borleis <jbe@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>