redirect command stderr to stdout
Redirect stderr to stdout for the actual autogen.sh, prepare, compile,
install and opkg-build command.
This preserves the chronological order of stdout vs. stderr when building
packages in parallel with output sync enabled.

As a side-effect, the stderr output of these commands won't show up on the
console for quiet builds. But most of that is not really useful by itself
anyways. And the resulting logfile is more readable now for the rest.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>