install_*: don't install a file at all if it's a symlink to /dev/null
This is mostly useful for install_alternative/install_alternative_tree.
There are use-cases where individual files installed by upstream packages
are unwanted for some reason.
With this change, such files can be masked by providing a symlink to
/dev/null as alternative. This makes it possible to remove such files
without copying and modifying the rule.
Note: this does not affect install_link, so creating links to /dev/null is
still possible.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>