ptxdist: implement image specific nfsroot
For the simple use-case (the rootfs contains all packages) the normal
nfsroot works well. However, this is not useful to work with root
filesystems that contain only a subset of all packages.

With this change, if <IMAGE>_NFSROOT can be set to 'YES' then
<platformdir>/nfsroot/<image> will be filled with the files from only the
packages that are part of this image.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>