ptxdist: add urlcheck command
Checking if an archive is available without downloading it is difficult and
not 100 percent reliable. For best result curl needs to fail for error http
status codes (--fail) and follow redirects (--location). A HEAD request
(--head) is used to avoid downloading the whole file, but we send a GET
anyways (--request GET) because some servers don't allow HEAD requests.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>