libmd: move to staging (license issues and no more users inside ptxdist)
libmd provides MD2, MD4, MD5, SHA-1 and RIPEMD-160 message digest algorithms.

Each of the respective implementations includes different license statements,
with wrappers under beer-ware license.
None of them is specified or extracted in the ptxdist rule.
In addition, to handle esp. the RSA license on MD2 and MD4, some build config
parameters should be introduced to be able to remove them.

In ptxdist, ntp was the only direct user of libmd, but provides its internal
MD5 implementation by now. This one is used if no libmd or similar is found.
From an incomplete check, libarchive can also make use of libmd, but prefers
OpenSSL/libcrypto, as being more complete.

Last, package seems to be not updated since 2009.

So instead of fixing all of the above, move to staging, so it can be removed
some day, unless somebody needs it and picks up above tasks.

Signed-off-by: Andreas Pretzsch <apr@cn-eng.de>
(mol: add staging comment to the help text)
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>