iproute2: update version and minor improvements
Update to iproute2 4.7.0, which - thanks to Alexander Aring - includes a
patch to fix a xtables-related build failure.

In addition we have to set TC_CONFIG_XT to y. Otherwise, the
hand-crafted Configure script enables legacy iptables code, which in
turn leads to multiple build failures.
Enabling the iptables package in ptxdist is still not required.

Also pass the kernel headers include directory to make.

Replace the description of tc with a few modified lines from the
manpage.

Signed-off-by: Clemens Gruber <clemens.gruber@pqgruber.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>