valgrind: add support for compressed debug-info
PTXdist separates the debug-info from the programs and libraries and
provides them in a .debug sub-directory near the programs and libraries
in a compressed manner (toolchain feature).
This change enables valgrind to still use the compressed debug-info to
improve its output logs.
Note: the new patch can be removed again on the next release of valgrind.

Signed-off-by: Juergen Borleis <jbe@pengutronix.de>