binutils: provide PIC code in the static libiberty
Many tools are using this static library all over the place. And some using
the 'libiberty' in their shared libraries ('latrace' for example) which
fails at link time because the 'binutils' pacakge provides a non-PIC
variant of this library only.
This change is a hack to force 'libiberty' to be compiled as a PIC object
to be able to link it against shared libraries as well.

Signed-off-by: Juergen Borleis <jbe@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>