gobject-introspection: Fix rpath lookup for symlinked packages
If the package directory is a symlink to a local_src directory, find by
default does not resolve the symlink. Therefore, libtool does not find
the .libs directories, which leads to linking errors for packages using
gobject-introspection (e.g., gst-plugins-bad1).

Add the -H command line option to resolve symlinks of the command line
arguments.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>