iputils: update s20121221 -> s20161105
The current version of iputils ping can handle IPv4 and IPv6 but
requires a crypto library dependency for ICMPv6 NI queries.
Add all possible choices: libgcrypt, nettle, openssl or none

tracepath works with IPv4 as well as IPv6 and it does not have to run as
root.
traceroute6 only works with IPv6 and requires superuser privileges.

Signed-off-by: Clemens Gruber <clemens.gruber@pqgruber.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>