image-rauc: new package
This adds a default image recipe for building a RAUC update Bundle out of the
systems rootfs.

In order to sign your update (mandatory) you need to place a valid certificate
and key file in your BSP. Their location is
default they are currently expected to be located at:

  $(PTXDIST_PLATFORMCONFIGDIR)/config/rauc/rauc.key.pem (key)
  $(PTXDIST_PLATFORMCONFIGDIR)/config/rauc/rauc.cert.pem (cert)

PTXdist will then create the bundle during a run of `ptxdist images`.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>