Debugging: add support for 'build-id's
Recent debugging tools support separate binaries and debug information
files. Providing a unique 'build-id' embedded in the binary enables the
debugging tools to select the correct and matching debug information
files from the system's root filesystem.

The 'build-id' is a 40 character calculated SHA1 from important sections in
the binary. The debug information files get stored at a central direcory at
"/usr/lib/debug/", with a special path based on the 'build-id'. If the
'build-id' is "123456789abcdef", the path and filename will be
".build-id/12/3456789abcdef.debug".

Its possible to have more than one 'build-id' in the binary. This
implementation prefers the first one in the list, which might be the
wrong one. It's unclear yet how the common debugging tools will behave in
this case.

Signed-off-by: Juergen Borleis <jbe@pengutronix.de>