ptxdist: handle symlinks correctly when comparing PTXDIST_TOPDIR and PTXDIST_WORKSPACE
PTXDIST_TOPDIR is set to based on the path to ptxdist with all symlinks
resolved. PTXDIST_WORKSPACE is set to the current working directory. If
this path contains symlinks then comparing the strings does not work as
expected.
Use '-ef' instead to check if both paths point to the same directory.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>