barebox_mlo: blacklist TARGET_BUILD_ID
We don't want build-id information in the barebox_mlo binaries.  It also
results in a build failure for Altera SOCFPGA with some toolchains that
end up outputting the ".note.gnu.build-id" section before the ".text"
section, which causes barebox's "scripts/fix_size" program to report
"invalid magic" in the file it is operating on and exit with a failure.

Fixes: 56b1e28fdbf78 ("Debugging: add support for 'build-id's")
Signed-off-by: Ian Abbott <abbotti@mev.co.uk>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>