libseccomp: install some useful development tools on demand
These tools are intended for development and thus required on demand only.
'scmp_app_inspector' is a shell script which requires some additional
tools from Busybox or Coreutils at run-time. Since both variants work,
it up to the user to select one variant. Which tools are required and
must be enabled in the menu is mentioned in the help text.

Signed-off-by: Juergen Borleis <jbe@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>