lighttpd: Rename memcache option to memcached
The memcached ./configure option for lighttpd changed, see the changeset
for upgrading lighttpd package from 1.4.39 to 1.4.45. For making the
configure option consistent with the menu entry, the menu variable is
renamed here. It seems lighttpd has switched from libmemcache to
libmemcached for memcached support.

Took this from a not applied patch from 2016, and added the necessary
migration helper suggested by Michael Olbrich, plus the dependency
changes in lighttpd.in.

However libmemcache and libmemcached are outdated and build against the
old libmemcached fails now, so mark it as broken.

Suggested-by: Andreas Geisenhainer <Andreas.Geisenhainer@atsonline.de>
Suggested-by: Michael Olbrich <m.olbrich@pengutronix.de>
Signed-off-by: Alexander Dahl <ada@thorsis.com>
[mol: LIGHTTPD_MOD_TRIGGER_B4_DL is also broken + commends for that]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>