iptables: bump version and fix large file support
The configure script handles the "--*-largefile" parameter badly. It
defaults to large file support if not given and always disables largefile
support if given (and it doesn't matter if 'enable' or 'disable' was used)

Due to the version bump, the previously required patch can be replaced,
because its already part of the new version.

Signed-off-by: Juergen Borleis <jbe@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>