poppler: version bump 0.18.0 -> 0.60.1
This is quite a version bump, spanning six years of poppler development.

Upstream now provides data for rendering cyrillic and CJK languages in
the package poppler-data, but this dependency is only detected at
compile time, so we make sure the build order is right. The build system
has changed from autotools to CMake.

Some applications require the old xpdf headers, the CPP wrapper or CMYK
support, so make these options selectable. Same for the Splash backend,
which can now also be built with single precision or fixed-point
arithmetic.

PNG and JPEG support is no longer broken; poppler now also includes an
internal DCT decoder, but it is officially unmaintained, so we leave it
out. Likewise, poppler allows using zlib to uncompress flate streams,
but with the warning that this is "not totally safe", so it is disabled
for now. The GTK dependency has been dropped as it is only necessary
for the documentation or some tests, which we don't build.

Additionally, poppler could benefit from packaging libopenjpeg and
libtiff in ptxdist for parsing JPEG2000 or TIFF embedded in PDFs.
Likewise, libnss could be packaged for PDF signature support. Also the
Qt5 wrapper is now selectable, but this is currently only build-tested
and needs to be tested in runtime.

Finally harmonize the wording of the kconfig options.

Signed-off-by: Roland Hieber <r.hieber@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>