at91bootstrap: Add patch to work with EABI toolchains
This one comes directly from the buildroot project and dates back to
2010. I had this on my BSP patch stack for at91bootstrap since 2012. I
only added a more or less matching subject for the patch, which is
missing in the patch buildroot has and set the time to the one of the
buildroot commit introducing the patch over there.

Also recreated whole patch series.

Signed-off-by: Alexander Dahl <ada@thorsis.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>