linuxptp: Update package and configs
Usualy we can not ensure only to use transparent clock switches.
Therefor we stick to the default E2E request-response mechanism.

The following options are highly depending on the usecase
so we remove them for the defaults.

path_trace_enabled
follow_up_info
tx_timestamp_timeout

With all other options being simple defaults the whole
config can be removed.

We also add some service files to start the ptp4l service on
dev eth0. In case the ptp4l daemon is configured
we add the option to install the phc2sys daemon to adjust the
system time by the ptp master.

Signed-off-by: Michael Grzeschik <m.grzeschik@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>