ptxd_make_world_common.sh: fix meson build for non-Debian distros
Debian patches glibc to add a special C.UTF-8 locale, which is not
available on most other distros.
For the meson build to work on other distros like ArchLinux, let's grep
the output of locale -a and try to find C.UTF-8, en_US.UTF-8 or an en_*
UTF-8 locale. If none can be found, fall back to any UTF-8 locale and
print a warning.

Signed-off-by: Clemens Gruber <clemens.gruber@pqgruber.com>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>