iproute2: cleanup rules
The pkg-config wrapper in ptxdist will only find packages that are selected
by the current ptxdist package. So the xtables, libelf libselinux and
libmnl checks work correctly. A local overwrite is not needed.

Also, the default WFLAGS are now ok and 'KERNEL_INCLUDE' is not used or
needed. 'iproute2' has it's own copy of all relevant headers.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>