wrapper: add rpath-link for some rpaths
With new versions of binutils, ld will prepend the sysroot to all absolute
rpaths before using them at build time. This breaks rpaths for build
directories.
Specifically it breaks building the gobject-introspection helper programs.

Fix this by adding the same path with -rpath-link.

Note: This is incomplete: It only works for "-Wl,-rpath,<path>" but not for
"-Wl,-rpath -Wl,<path>". It fixes all known issues in PTXdist so it is good
enough for now.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>