crda: extend rule to download versioned regulatory.bin
The crda reads the regulatory database information from regulatory.bin,
which is updated upon regulatory changes and is versioned.

Extend the crda.make to download a specified version of the database to
document the used database version. This allows to drop the copy of the
regulatory.bin from the projectroot.

Still allow to override the database in a BSP projectroot if necessary.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
[mol: simplify install stage]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>