ptxdist: find config files in multiple layers
This is the first step to support layers in PTXdist.

A layer is just a subdirectory (or symlink) named 'base' in the BSP
workspace. Each layer looks just like another BSP. The first layer is
implicitly or explicitly PTXdist itself.

Automatic config file detection works as before in the last layer. If no
config file is found, then the previous tlayer is checked the same way, and so
on.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>