m4/coreutil: add gnulib patches for GLibc-2.28
This adds the necessary patches for m4 and coreutils to compile with glibc-2.28.
These can be removed when new releases are available.

Both m4 and coreutils use an in-tree copy of gnulib which abstracts some of the
more ugly OS differences. To support the new glibc only changes in gnulib are
necessary, so the patches for both packages are nearly identical.

This version is manually applied for both packages from
4af4a4a71827c0bc5e0ec67af23edef4f15cee8e
in the upstream gnulib repository.

v2:
- fix spelling mistakes
- mention the in-tree gnulib copy to justify the single patch
from Uwe Kleine-König
- Integrate original commit message, author and date
- add backport note and my SoB
from Denis Osterland

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>