weston: install weston.ini into /etc
Weston brings an example weston.ini file, which is not installed by
default. Use the shipped template to generate a weston.ini for the
target system and install it to the default system global path.

Allow to disable weston.ini to use the build-in default configuration or
to override it in the projectroot.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>