rules: python3-shiboken: Fix recursive path replacement in cmake file
The install.post target replaces matching patterns in paths in the
generated cmake file. If a replacement somehow matches the next script's
regexp, the paths are malformed.

By adding the '@' at the beginning of each replacement, it is assured the
next pattern will never match an earlier replacement. At last, '"@' is
replaced with '"', removing the added '@'.

In our particular case, the workdir is located in /usr/src/, which matched
the last script's regexp, replacing "/usr with "$(SYSROOT).

Signed-off-by: Pim Klanke <pim@protonic.nl>
[mol: just one '-i' at the beginning]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>