ptxd_print_path: handle layers correctly
'oldconfig' is executed recursivly. Each time, the last layer is removed
from PTXDIST_PATH. As a result, ptxd_print_path will produce paths starting
with 'base/' in the current layer.
Also, any platformconfigdir or layer paths are not used anyways, because
the workspace matches first.
So just use the workspace and topdir explicitly.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>