ptxd_install_replace_figlet: Use sed instead of awk to quote backslashes
TL;DR: awk and backslashes in the replacement argument of gsub is ugly.

With GNU Awk 4.0.2 we have:

	$ echo '\' | awk '{ gsub("\\\\", "\\\\\\\\\\\\\\\\"); print }'
	\\\\\\\\

and with GNU Awk 4.1.4 (and newer) we have:

	$ echo '\' | awk '{ gsub("\\\\", "\\\\\\\\\\\\\\\\"); print }'
	\\\\

This results in too much backslashes in /etc/issues for users of the
former version.

For the glory details I recommend reading through

	https://www.gnu.org/software/gawk/manual/html_node/Gory-Details.html
	https://bugs.debian.org/909147

As it's awkward to come up with an expression that does the same on both
versions (e.g. using --posix) use sed instead.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>