ptxd_lib_dgen: make sure $(<PKG>_SOURCE) refers to the primary package
To get the package for a source archive we define for each package:

$(<PKG>_SOURCE) := <PKG>

With host and cross packages that share the source with the target package,
the same variable may be defined multiple times. Which one 'wins' is
undefined.
Avoid this by only defining the variable for the target package.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>