rauc: fix quoting for when replacing @RAUC_BUNDLE_COMPATIBLE@
The default for RAUC_COMPATIBLE is
"${PTXCONF_PROJECT_VENDOR} ${PTXCONF_PROJECT}". As a result
PTXCONF_RAUC_COMPATIBLE evaluates to something like this:
""<the-vendor>" "<the-project>""
So the space is quoted but not the rest, so install_replace will fail if
PROJECT_VENDOR or PROJECT contain a space.

Fix the quoting to avoid this.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>