wrapper: try to make all debug paths relative to PTXDIST_WORKSPACE
The debug sections containing these paths are used to calculate the
build-id, so building a BSP in a different directory will result in a
different build-id even if nothing else changes.
Avoid this by using -fdebug-prefix-map to make the paths relative.
Also expand the gdb wrapper to handle this even if it is called from a
different directory.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>