bzip2: fix URL (bzip.org archive gone, use Gentoo as mirror)
As of 22 September 2018, bzip2 sources are not longer available
on http://www.bzip.org/. They state "Find the latest version on
SourceForge", just I cannot find it there. Only as part of some
other projects, which would be unreliable.

So for now, refer to Gentoo source archives, i.e.
http://distfiles.gentoo.org/distfiles/bzip2-1.0.6.tar.gz

This should be revised as soon as there is another (official)
source location.

Signed-off-by: Andreas Pretzsch <apr@cn-eng.de>
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>