ptxd_lib_dgen: let packages depend on config files
For kconfig based packages, the prepare stage should depend on the config
file to ensure that the package is rebuilt when the config changes.
Just do it for all packages. It's a noop if no config is defined.
Also add the equivalent dependency for image packages.

Only add dependencies if <PKG>_CONFIG seems to be an absolute path. This is
necessary to avoid problems with genimage packages: The config file is a
relative path to config/images/ and a dependency is generated at
built-time.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>