ptxd_make_world_init: improve logfile ordering
If output-sync is enabled, then all stdout and stderr is collected for a
target and dumped at the end. stdout and stderr are already merged to
achieve proper interleaving.
However anything written directly to the logfile comes first.
Redirect the 'logfile only' file descriptor to stdout if quiet is enabled.
The result is the same but correctly interleaved.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>