ptxdist: fix PTXDIST_TOPDIR as explicit last layer
If PTXdist is an explicit last layer, then it is often a symlink.
PTXDIST_TOPDIR is always resolved with 'realpath', so do the same when
setting up the layers.
Make sure that the last path in PTXDIST_LAYERS is exactly PTXDIST_TOPDIR.
This is needed used in other places (e.g. ptxd_kconfig_update()).

This also fixes calling ptxdist from PTXDIST_TOPDIR when
$PWD != `realpath $PWD`.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>