ptxd_lib_dgen: make sure the source -> package mapping is defined
For all active packages "$(<PKG>_SOURCE): <PKG>" is defined. Host and cross
packages are skipped if the target package has the same source. As a
result, the mapping is missing if the host package is enabled but not the
target package.
Fix this by defining the mapping for host and cross packages if the target
package is disabled.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>